[FIX] Import microdrop-plugin-manager at runtime
Prior to this commit, `mpm` (i.e., Microdrop plugin manager) is imported
in the `init_hooks` and `create_plugin` modules on import, making
`microdrop_plugin_manager` a required package to import `init_hooks` or
`create_plugin`.  However, building docs on [`readthedocs`][1] fails
when `microdrop_plugin_manager` is a required package, since `pywin32`
is a required package for `microdrop_plugin_manager`, but `pywin32` is
not available on the Python Package Index.

As of this commit, `mpm` is only imported in the functions that actually
use the import, ensuring `mpm` is only imported at runtime.  This
permits building docs on [`readthedocs`][1].

[1]: https://readthedocs.org